+++ backend/src/main/java/com/capstone/grocery/repository/GroceryStoreRepository.java
@@ -6,4 +6,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroceryStoreRepository extends JpaRepository<GroceryStore, Long> {}
+public interface GroceryStoreRepository extends JpaRepository<GroceryStore, Integer> {}
